+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Tags can be assigned to batches
 - Filter batches by tag
 - Add backend_configuration arg to create_batch for serialised configurations
+- New MockSDK/MockServer, used for testing code that uses this SDK
 
 ## [0.20.2]